Fixed Parse queries, connected views, and cleaned up code.
Placed all Parse queries where they belong in the proper cells and view
controllers. There is only a single Parse query made at the very
beginning, and all subsequent data is mined from that single object and
is passed as needed. Images are loaded only when necessary, and code is
commented. Labels are gender-specific. Looks OK in iPhone 6, but best
in iPhone 5.