added some commenting functionality
The backend is configured properly to handle commenting, and newsfeed
updates populate with data returned through relational pointers. It is
currently possible to click on a newsfeed update and create a comment
saved to Parse and linked with the logged in user. Not yet able to
display queried comments properly.